Make LogSink::Severity a standalone enum class Severity.
This makes code that doesn't use one of the high-level functions
like LogError more elegant: Log(LogSink::Severity::ERROR)
turns into Log(Severity::ERROR).

The reason to not use LogError is that it mangles the output,
which is undesirable when the output will be produced in many chunks,
or across many lines.

Signed-off-by: Andrew Zonenberg <azonenberg@drawersteak.com>